check the confidence ratio

+++ src/evaluation.py
@@ -19,10 +19,10 @@ def prediction(loader, model, args, verbose=False):
     raw_pred = nn.Softmax(dim=1)(pred)
     _, pred = torch.max(raw_pred, 1)
 
-    return raw_pred.detach().cpu(), pred.detach().cpu(), true.cpu()
+    return raw_pred.detach().cpu(), pred.detach().cpu(), true.cpu().int()
 
 def cal_acc(loader, model, args, verbose=False):
     _, pred, true = prediction(loader, model, args, verbose)
-    acc = (pred.float() == true).float().mean()
+    acc = (pred == true).float().mean()
     return acc.item()
